chore: move weak test connection to global
