+++ test/esm/global/test-pool-connect-error.test.mts
@@ -5,10 +5,10 @@ import mysql from '../../../index.js';
 // The process is not terminated in Deno
 if (typeof Deno !== 'undefined') process.exit(0);
 
-process.on('uncaughtException', (err: NodeJS.ErrnoException) => {
-  if (err.code === 'PROTOCOL_CONNECTION_LOST' || err.code === 'EPIPE') return;
-  throw err;
-});
+// process.on('uncaughtException', (err: NodeJS.ErrnoException) => {
+//   if (err.code === 'PROTOCOL_CONNECTION_LOST' || err.code === 'EPIPE') return;
+//   throw err;
+// });
 
 await describe('Pool Connect Error', async () => {
   await it('should emit error code 1040 for connection and pool', async () => {
